Input sanitization

+++ transform_number.h
@@ -1,6 +1,6 @@
 #ifndef TRANSFORM_NUMBER_H
 #define TRANSFORM_NUMBER_H
 
-void transform(int num);
+void transform(long num);
 
 #endif
